[SC-159] Install dependencies with aptdcon to avoid dpkg lock contention
Uses aptdcon for installing packages to avoid apt lock contention.
Inspired by [this AskUbuntu question](https://askubuntu.com/questions/132059/how-to-make-a-package-manager-wait-if-another-instance-of-apt-is-running).

aptdcon submits requests to aptdaemon which queues them, so our
request will wait instead of failing when it can't acquire a lock

Updates `SourceImage` to packer-base-ubuntu-bionic v1.0.8
(which installs aptdaemon package)