Single volume instance
Setting device name to "/dev/xvda" causes AWS to add an additional
volume to the instance because xvda is for awslinux.   We only want
a single root volume so we can tag it for cost tracking.
For Ubuntu the device name should be "/dev/sda1".

Also the default volume size can be a little bigger.